requirements.txt
jupyter
torch
torchvision
Pillow
tqdm
kaggle
scikit-learn
